Kai/dev

+++ Backend/controller/SponsorController.js
@@ -1,14 +1,33 @@
 const express = require("express");
 const router = express.Router();
 const service = require("../service/SponsorService");
+const logoService = require("../service/LogoService");
 const { NOT_FOUND, OK } = require("./Constants");
+const { query } = require("express-validator");
 const { authToken, requesterIsNotModerator } = require("./Auth");
 const {
   checkStringObjectNotEmpty,
   checkOptionalStringObjectNotEmpty,
   checkOptionalIsBoolean,
+  checkIsUuidObject,
+  checkParamIsUuid,
   validationHasErrors,
 } = require("./Validation");
+const multer = require("multer");
+const path = require("path");
+
+const IMAGE_STORE = "data/images/sponsors";
+
+const storage = multer.diskStorage({
+  destination: function (req, file, cb) {
+    cb(null, IMAGE_STORE);
+  },
+  filename: function (req, file, cb) {
+    const prefix = Date.now();
+    cb(null, prefix + "-" + file.originalname);
+  },
+});
+const imageUpload = multer({ storage: storage });
 
 // @desc Gets all sponsors
 // @route GET /sponsors/
@@ -24,6 +43,75 @@ router.get("/", authToken, async (req, res, next) => {
     next(error);
   }
 });
+
+// @desc Gets the logo of an sponsor
+// @route GET /media/logo/:id
+
+router.get(
+  "/logo/:id",
+  checkParamIsUuid("id"),
+  query("thumb").optional().isBoolean(),
+  async (req, res, next) => {
+    if (validationHasErrors(req, res)) return;
+    const id = req.params.id;
+    const thumb = req.query.thumb;
+
+    try {
+      const logo = await logoService.getById(id);
+
+      if (!logo) return res.sendStatus(NOT_FOUND);
+
+      const fullfilepath = thumb
+        ? path.join(path.resolve(), logo.pathThumb)
+        : path.join(path.resolve(), logo.path);
+
+      return res.type(logo.mimetype).sendFile(fullfilepath);
+    } catch (error) {
+      next(error);
+    }
+  }
+);
+
+// @desc Uploads a new logo for a sponsor
+// @route POST /sponsors/logo
+// @access Only moderator
+
+router.post(
+  "/logo",
+  authToken,
+  imageUpload.single("image"),
+  checkIsUuidObject("sponsorId"),
+  async (req, res, next) => {
+    try {
+      if (await requesterIsNotModerator(req, res)) return;
+
+      const originalname = req.file.originalname;
+      const mimetype = req.file.mimetype;
+      const size = req.file.size;
+      const path = req.file.path;
+      const sponsorId = req.body.sponsorId;
+
+      const sponsor = await service.getById(sponsorId);
+      if (!sponsor) return res.sendStatus(NOT_FOUND);
+
+      if (sponsor.Logo) {
+        logoService.deleteOldLogo(sponsor);
+      }
+
+      const logo = await logoService.create({
+        originalname,
+        mimetype,
+        size,
+        path,
+        sponsorId,
+      });
+
+      res.json(logo);
+    } catch (error) {
+      next(error);
+    }
+  }
+);
 
 // @desc Adds a sponsor to the db
 // @route POST /sponsors/
@@ -81,6 +169,7 @@ router.post(
 router.put(
   "/:id",
   authToken,
+  checkParamIsUuid("id"),
   checkOptionalStringObjectNotEmpty("name"),
   checkOptionalStringObjectNotEmpty("type"),
   checkOptionalStringObjectNotEmpty("website"),
@@ -91,11 +180,12 @@ router.put(
   checkOptionalIsBoolean("isGoldSponsor"),
   async (req, res, next) => {
     if (validationHasErrors(req, res)) return;
+    const id = req.params.id;
 
     try {
       if (await requesterIsNotModerator(req, res)) return;
 
-      const sponsor = await service.getById(req.params.id);
+      const sponsor = await service.getById(id);
 
       sponsor.name = req.body.name ?? sponsor.name;
       sponsor.type = req.body.type ?? sponsor.type;
@@ -139,20 +229,27 @@ router.put(
 // @route DELETE /sponsors/:id
 // @access Only owner
 
-router.delete("/:id", authToken, async (req, res, next) => {
-  const id = req.params.id;
-  try {
-    if (await requesterIsNotModerator(req, res)) return;
+router.delete(
+  "/:id",
+  checkParamIsUuid("id"),
+  authToken,
+  async (req, res, next) => {
+    if (validationHasErrors(req, res)) return;
+    const id = req.params.id;
 
-    const sponsor = await service.getById(id);
+    try {
+      if (await requesterIsNotModerator(req, res)) return;
 
-    if (!sponsor) return res.sendStatus(NOT_FOUND);
+      const sponsor = await service.getById(id);
 
-    const numberOfDestroyedRows = await service.delete(id);
-    res.json(numberOfDestroyedRows);
-  } catch (error) {
-    next(error);
+      if (!sponsor) return res.sendStatus(NOT_FOUND);
+
+      const numberOfDestroyedRows = await service.delete(id);
+      res.json(numberOfDestroyedRows);
+    } catch (error) {
+      next(error);
+    }
   }
-});
+);
 
 module.exports = router;
